Set spacing around display math
Setting spacing around display math is a bit tricky. The obvious

    \setupformulas[spacebefore=small, spaceafter=small]

does not work. I think that this is because I use

    \setupwhitespace[big]

and ConTeXt vertical spacing model is such that the largest vspace value (big
in this case) wins.

After a bit of trial and error, I settled on this:

    \setupformula
        [spacebefore={nowhite,quarterline}, spaceafter={back,nowhite,quarterline}]

For `spacebefore`, the parameter `nowhite` kills all the whitespace and
then `quarterline` sets the space before the formula to be 0.25\lineheight.

Surprisingly, the same setting `{nowhite,quarterline}` does not work when used
for `spaceafter`. We need an addition `back` parameter to kill the previous
vpsace.